+++ Runtime/D3D12RHI/D3D12RHIShader.h
@@ -94,7 +94,7 @@ public:
 
     FORCEINLINE SD3D12ShaderParameter GetConstantBufferParameter( uint32 ParameterIndex )
     {
-        D3D12_ERROR( ParameterIndex < ConstantBufferParameters.Size(), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(ConstantBufferParameters.Size()) + " slots");
+        D3D12_ERROR( ParameterIndex < static_cast<uint32>(ConstantBufferParameters.Size()), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(ConstantBufferParameters.Size()) + " slots");
         return ConstantBufferParameters[ParameterIndex];
     }
 
@@ -105,7 +105,7 @@ public:
 
     FORCEINLINE SD3D12ShaderParameter GetShaderResourceParameter( uint32 ParameterIndex )
     {
-        D3D12_ERROR( ParameterIndex < ShaderResourceParameters.Size(), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(ShaderResourceParameters.Size()) + " slots");
+        D3D12_ERROR( ParameterIndex < static_cast<uint32>(ShaderResourceParameters.Size()), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(ShaderResourceParameters.Size()) + " slots");
         return ShaderResourceParameters[ParameterIndex];
     }
 
@@ -116,7 +116,7 @@ public:
 
     FORCEINLINE SD3D12ShaderParameter GetUnorderedAccessParameter( uint32 ParameterIndex )
     {
-        D3D12_ERROR( ParameterIndex < UnorderedAccessParameters.Size(), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(UnorderedAccessParameters.Size()) + " slots");
+        D3D12_ERROR( ParameterIndex < static_cast<uint32>(UnorderedAccessParameters.Size()), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(UnorderedAccessParameters.Size()) + " slots");
         return UnorderedAccessParameters[ParameterIndex];
     }
 
@@ -127,7 +127,7 @@ public:
 
     FORCEINLINE SD3D12ShaderParameter GetSamplerStateParameter( uint32 ParameterIndex )
     {
-        D3D12_ERROR( ParameterIndex < SamplerParameters.Size(), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(SamplerParameters.Size()) + " slots");
+        D3D12_ERROR( ParameterIndex < static_cast<uint32>(SamplerParameters.Size()), "Trying to access ParameterIndex=" + ToString(ParameterIndex) + ", but the shader only has " + ToString(SamplerParameters.Size()) + " slots");
         return SamplerParameters[ParameterIndex];
     }
 
@@ -159,6 +159,7 @@ public:
     static bool GetShaderReflection( class CD3D12BaseShader* Shader );
 
 protected:
+
     template<typename TD3D12ReflectionInterface>
     static bool GetShaderResourceBindings( TD3D12ReflectionInterface* Reflection, CD3D12BaseShader* Shader, uint32 NumBoundResources );
